Reworked Keras interface.

+++ interface/pytorch/datasets_test.py
@@ -9,37 +9,12 @@ from lib.utils import testUtils
 from interface.pytorch import datasets
 
 
-class TestHelperFunctions(testUtils.Test_carsDb):
-    def test_used_keys(self):
+class TestHelperFunctions(unittest.TestCase):
+    def test_filterKeys(self):
         sample = {'image': np.zeros(3), 'objectid': 1}
-        sample = datasets._filter_keys(['image'], sample)
+        sample = datasets._filterKeys(['image'], sample)
         self.assertEqual(len(sample), 1)  # only "image" key is left.
         self.assertTrue('image' in sample)
-
-    def test_apply_transform_callable(self):
-        transform_group = lambda x: x['image']
-        sample = {'image': np.zeros(3), 'objectid': 1}
-        sample = datasets._apply_transform(transform_group, sample)
-        self.assertTrue(isinstance(sample, (np.ndarray, np.generic)))
-
-    def test_apply_transform_list(self):
-        transform_group = [lambda x: x['image'], lambda x: x['objectid']]
-        sample = {'image': np.zeros(3), 'objectid': 1, 'name': 'car'}
-        sample = datasets._apply_transform(transform_group, sample)
-        self.assertTrue(isinstance(sample, list))
-        self.assertEqual(len(sample), 2)  # only "image" and 'objectid' left.
-        self.assertTrue(isinstance(sample[0], (np.ndarray, np.generic)))
-        self.assertTrue(isinstance(sample[1], int))
-
-    def test_apply_transform_dict(self):
-        transform_group = {
-            'image': lambda x: x[:, :, 0],
-            'name': lambda _: 'hi'
-        }
-        sample = {'image': np.zeros((10, 10, 3)), 'objectid': 1, 'name': 'car'}
-        sample = datasets._apply_transform(transform_group, sample)
-        self.assertEqual(len(sample['image'].shape), 2)  # Grayscale image.
-        self.assertEqual(sample['name'], 'hi')  # All names replaced to "hi".
 
 
 class TestImageDataset(testUtils.Test_carsDb):
@@ -48,7 +23,8 @@ class TestImageDataset(testUtils.Test_carsDb):
         shutil.copy(testUtils.Test_carsDb.CARS_DB_PATH, self.tmp_in_db_file)
 
     def tearDown(self):
-        os.remove(self.tmp_in_db_file)
+        if os.path.exists(self.tmp_in_db_file):
+            os.remove(self.tmp_in_db_file)
 
     def test_general(self):
         dataset = datasets.ImageDataset(
@@ -75,6 +51,18 @@ class TestImageDataset(testUtils.Test_carsDb):
         self.assertTrue('height' in object_sample)
         self.assertTrue('name' in object_sample)
         self.assertTrue('score' in object_sample)
+
+    def test_used_keys(self):
+        dataset = datasets.ImageDataset(
+            testUtils.Test_carsDb.CARS_DB_PATH,
+            rootdir=testUtils.Test_carsDb.CARS_DB_ROOTDIR,
+            used_keys=['mask', 'score'])
+        self.assertEqual(len(dataset), 3)  # 3 images.
+        sample = dataset[0]
+        self.assertTrue(isinstance(sample, dict))
+        self.assertEqual(len(sample), 2)
+        self.assertTrue('mask' in sample)
+        self.assertTrue('score' in sample)
 
     def test_where_image(self):
         dataset = datasets.ImageDataset(
@@ -108,7 +96,8 @@ class TestObjectDataset(testUtils.Test_carsDb):
         shutil.copy(testUtils.Test_carsDb.CARS_DB_PATH, self.tmp_in_db_file)
 
     def tearDown(self):
-        os.remove(self.tmp_in_db_file)
+        if os.path.exists(self.tmp_in_db_file):
+            os.remove(self.tmp_in_db_file)
 
     def test_general(self):
         dataset = datasets.ObjectDataset(
@@ -132,6 +121,18 @@ class TestObjectDataset(testUtils.Test_carsDb):
             'SELECT value FROM properties WHERE objectid=? AND key="result"',
             (sample['objectid'], ))
         self.assertEqual(dataset.c.fetchall(), [("0.5", )])
+
+    def test_used_keys(self):
+        dataset = datasets.ObjectDataset(
+            testUtils.Test_carsDb.CARS_DB_PATH,
+            rootdir=testUtils.Test_carsDb.CARS_DB_ROOTDIR,
+            used_keys=['mask', 'score'])
+        self.assertEqual(len(dataset), 3)  # 3 objects.
+        sample = dataset[0]
+        self.assertTrue(isinstance(sample, dict))
+        self.assertEqual(len(sample), 2)
+        self.assertTrue('mask' in sample)
+        self.assertTrue('score' in sample)
 
     def test_where_object(self):
         # All objects should be "cars".
